+++ meetup2apricot/event_registration_type.py
@@ -26,11 +26,11 @@ class EventRegistrationType:
             "GuestPrice": self.price,
             "Availability": "Everyone",
             "MaximumRegistrantsCount": self.maximum_registrants_count,
-            "GuestRegistrationPolicy": "CollectContactDetails",
+            "GuestRegistrationPolicy": "NumberOfGuests",
             "UnavailabilityPolicy": "ShowDisabled",
             "CancellationBehaviour": "AllowUpToPeriodBeforeEvent",
             "CancellationDaysBeforeEvent": 2,
-            "IsWaitlistEnabled": False
+            "IsWaitlistEnabled": True
             }