+++ tests/sample_apricot_json.py
@@ -31,7 +31,16 @@ EXPECTED_FREE_EVENT_JSON = {
         "AccessControl": {
             "AccessLevel": "Public"
             },
-        "PaymentMethod": "OnlineOnly"
+        "PaymentMethod": "OnlineOnly",
+        "IsWaitlistEnabled": True,
+        "WaitlistSettings": {
+            "WaitlistType": "Manual",
+            "InformationToCollect": "ContactInformation"
+            },
+        "AttendeesDisplaySettings": {
+                "VisibleTo": "Members",
+                "ShowPendingAttendees": False
+            }
         }
     }