+++ HISTORY.rst
@@ -15,6 +15,14 @@ Recent History
 .. Next Release
 .. ------------------
 
+1.0.0 (2020-11-08)
+------------------
+
+* Turn on waitlist with manual processing.
+* Collect guest counts, but no personal information.
+* Throttle Wild Apricot requests.
+* Allow members to see the list of event registrants.
+
 0.2.0 (2020-11-08)
 ------------------